requirements.txt
asgiref==3.4.1
Django==3.2.6
django-bootstrap-datepicker-plus==3.0.5
django-cleanup==5.2.0
django-widget-tweaks==1.4.8
pytz==2021.1
sqlparse==0.4.1
